+++ apps/astarte_pairing/config/test.exs
@@ -98,10 +98,10 @@ config :astarte_pairing,
 config :astarte_pairing, :astarte_instance_id, "test"
 
 config :astarte_pairing, :enable_fdo, true
-config :astarte_pairing, :base_url_domain, "api.astarte.localhost"
+config :astarte_pairing, :base_url_host, "api.astarte.localhost"
 config :astarte_pairing, :base_url_port, 4003
 config :astarte_pairing, :base_url_protocol, :http
-config :astarte_fdo, :base_url_domain, "api.astarte.localhost"
+config :astarte_fdo, :base_url_host, "api.astarte.localhost"
 config :astarte_fdo, :base_url_port, 4003
 config :astarte_fdo, :base_url_protocol, :http
 config :astarte_pairing, :enable_credential_reuse, true