+++ libs/astarte_fdo/test/astarte_fdo/config_test.exs
@@ -22,6 +22,7 @@ defmodule Astarte.FDO.ConfigTest do
   import Mimic
 
   alias Astarte.FDO.Config
+  alias Astarte.FDO.Config.BaseURLHost
   alias Astarte.FDO.Config.BaseURLProtocol
 
   describe "BaseURLProtocol.cast/1" do
@@ -47,12 +48,22 @@ defmodule Astarte.FDO.ConfigTest do
   end
 
   describe "base_url!/0" do
-    test "builds the URL from protocol, domain and port" do
+    test "builds the URL from protocol, a domain host and port" do
+      domain_host = %BaseURLHost{type: :domain, value: "astarte.example.com"}
       stub(Config, :base_url_protocol!, fn -> :https end)
-      stub(Config, :base_url_domain!, fn -> "astarte.example.com" end)
+      stub(Config, :base_url_host!, fn -> domain_host end)
       stub(Config, :base_url_port!, fn -> 443 end)
 
       assert Config.base_url!() == "https://astarte.example.com:443"
+    end
+
+    test "builds the URL from an IP host" do
+      ip_host = %BaseURLHost{type: :ip, value: {192, 168, 1, 10}}
+      stub(Config, :base_url_protocol!, fn -> :http end)
+      stub(Config, :base_url_host!, fn -> ip_host end)
+      stub(Config, :base_url_port!, fn -> 4003 end)
+
+      assert Config.base_url!() == "http://192.168.1.10:4003"
     end
   end
 end